GUIEvent: added key for "multislider"
The old GUIEvent used the key “scMultislider”, which is out-dated.  I
kept this key, but duplicated it and changed the name to simply
“multislider.”

I updated the GUIEvent help file to reflect this.

Using \multislider in a ConductorGUI still doesn’t work, but it didn’t
before and at least the naming’s consistent now.  I’m hoping to fix
\multislider next.